Added safety checking for empty ID.

+++ workarounds/livecd-post-reboot/CASMINST-1309/CASMINST-1309.sh
@@ -21,7 +21,14 @@ for bmc in $bmcs; do
   bad_ips=$(curl -s -k -H "Authorization: Bearer ${TOKEN}" "https://api-gw-service-nmn.local/apis/smd/hsm/v2/Inventory/EthernetInterfaces?ComponentID=$bmc" | jq -r '.[] | ."IPAddresses" | .[] | ."IPAddress"')
   for ip in $bad_ips; do
     id=$(curl -s -k -H "Authorization: Bearer ${TOKEN}" "https://api-gw-service-nmn.local/apis/smd/hsm/v2/Inventory/EthernetInterfaces?IPAddress=$ip" | jq -r '.[] | ."ID"')
-    echo "Deleting $id from EthernetInterfaces"...
-    curl -X DELETE -s -k -H "Authorization: Bearer ${TOKEN}" "https://api-gw-service-nmn.local/apis/smd/hsm/v2/Inventory/EthernetInterfaces/$id" | jq
+
+    # Make sure ID isn't blank.
+    if [ -z "$id" ]
+    then
+      echo "$id blank when trying to find an owner for IP $ip!"
+    else
+      echo "Deleting $id from EthernetInterfaces"...
+      curl -X DELETE -s -k -H "Authorization: Bearer ${TOKEN}" "https://api-gw-service-nmn.local/apis/smd/hsm/v2/Inventory/EthernetInterfaces/$id" | jq
+    fi
   done
 done
